[refactor] playground client 패키지 변경 (#341)
